@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start" : "nodemon server/server.js",
+    "start": "nodemon server/server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "webpack -d --watch"
   },
@@ -18,12 +18,18 @@
     "url": "https://github.com/HR-Butternut/Catwalk/issues"
   },
   "devDependencies": {
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.0",
+    "@babel/preset-react": "^7.16.0",
     "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
+    "babel-jest": "^27.3.1",
+    "babel-loader": "^8.2.3",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
+    "css-loader": "^5.2.7",
     "jest": "^27.3.1",
-    "webpack": "^2.2.1",
+    "style-loader": "^2.0.0",
+    "webpack": "^4.46.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.4.0"
   },
@@ -32,8 +38,13 @@
     "axios": "^0.24.0",
     "babel": "^6.23.0",
     "bootstrap": "^5.1.3",
+    "jquery": "^3.6.0",
     "path": "^0.12.7",
+    "popper.js": "^1.16.1",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-bootstrap": "^2.0.2",
+    "react-dom": "^17.0.2",
+    "react-progressbar": "^15.4.1",
+    "react-star-ratings": "^2.3.0"
   }
 }
