@@ -1,3 +1,6 @@
+import 'bootstrap/dist/css/bootstrap.min.css';
+import $ from 'jquery';
+import Popper from 'popper.js';
 import React, {useState, useEffect} from 'react';
 import ReactDOM from 'react-dom';
 import axios from 'axios';
@@ -5,7 +8,6 @@ import Overview from './overview/index.jsx';
 import Questions from './questions/index.jsx';
 import Reviews from './reviews/index.jsx';
 import Product_Id_Context from './context.jsx';
-
 
 function App() {
   var [product_id, setProduct_id] = useState(0);
@@ -35,7 +37,7 @@ function App() {
   }
 
   /* Use Effects second paramenter makes sure that it only runs when the app is mounted like component did mount */
-  
+
   useEffect(() => {
     singleItemRequest();
   }, [])
@@ -45,7 +47,7 @@ function App() {
       <Product_Id_Context.Provider value={product_id}>
         <Overview/>
         <Questions/>
-        <Reviews/>
+        {/* <Reviews/> */}
       </Product_Id_Context.Provider>
     </div>
   )
